Add support for meta resources
This commit adds support for meta resources: resources that have no
content and return a non-success status. Mostly this means resources
that have moved (and should return 301) or resources that have been
deleted (and should return 410 or 404).

This feature is available by creating a simple text file called META in
the root directory of websites. The file contains directions
line-by-line, and each line is in the following format:

<SLUG> <STATUS> [MOVED_TO]

Where SLUG is the resource's path, STATUS is one of "permanent", "gone"
or "not_found", and MOVED_TO is the path to where the resource has moved
(only if status is "permanent").

The server will properly handle such resources and issue a response with
the correct status (and Location header if appropriate).